# openapi.yaml
openapi: 3.1.0
info:
  title: HTMLTrust Directory API
  description: |
    Reference implementation of the HTMLTrust trust directory API. Provides endpoints
    for author management, content signing and verification, claims, and a federated
    trust directory with reputation tracking.
  version: 1.0.0
  contact:
    name: HTMLTrust
    url: https://github.com/ArcadeLabsInc/htmltrust-server-reference
    email: jason@jason-grey.com
servers:
  - url: https://api.contentsigning.example.com/v1
    description: Production server
  - url: https://api.staging.contentsigning.example.com/v1
    description: Staging server
...
security:
  - {} # No default security
components:
  securitySchemes:
    AuthorApiKey:
      type: apiKey
      in: header
      name: X-AUTHOR-API-KEY
      description: API key linked to a specific author for author-specific operations

    GeneralApiKey:
      type: apiKey
      in: header
      name: X-API-KEY
      description: General API key for authenticated operations

    AdminApiKey:
      type: apiKey
      in: header
      name: X-ADMIN-API-KEY
      description: Admin-level API key for administrative operations
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      example:
        code: "UNAUTHORIZED"
        message: "Invalid API key"
    Author:
      type: object
      required:
        - id
        - name
        - keyType
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the author
        name:
          type: string
          description: Name of the author
        description:
          type: string
          description: Description of the author
        url:
          type: string
          format: uri
          description: URL associated with the author
        keyType:
          type: string
          enum: [HUMAN, AI, HUMAN_AI_MIX, ORGANIZATION]
          description: Type of the author key
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        name: "John Doe"
        description: "Technology writer and researcher"
        url: "https://johndoe.example.com"
        keyType: "HUMAN"
        createdAt: "2023-01-01T12:00:00Z"
        updatedAt: "2023-01-01T12:00:00Z"
    PublicKey:
      type: object
      required:
        - id
        - authorId
        - key
        - algorithm
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the public key
        authorId:
          type: string
          format: uuid
          description: ID of the author this key belongs to
        key:
          type: string
          description: The public key in PEM format
        algorithm:
          type: string
          description: The cryptographic algorithm used
          enum: [RSA, ECDSA, ED25519]
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp (if applicable)
      example:
        id: "123e4567-e89b-12d3-a456-426614174001"
        authorId: "123e4567-e89b-12d3-a456-426614174000"
        key: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA...\n-----END PUBLIC KEY-----"
        algorithm: "RSA"
        createdAt: "2023-01-01T12:00:00Z"
        expiresAt: "2024-01-01T12:00:00Z"
    Claim:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the claim type
        name:
          type: string
          description: Name of the claim type
        description:
          type: string
          description: Description of what this claim type represents
        possibleValues:
          type: array
          description: Possible values for this claim type (if applicable)
          items:
            type: string
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        id: "123e4567-e89b-12d3-a456-426614174002"
        name: "ContentType"
        description: "The type of content being signed"
        possibleValues: ["Article", "Image", "Video", "Audio", "Dataset"]
        createdAt: "2023-01-01T12:00:00Z"
        updatedAt: "2023-01-01T12:00:00Z"
...
paths:
  /authors:
    post:
      tags:
        - Authors
      summary: Create a new author and key pair
      description: Creates a new author profile and generates a key pair. The private key is stored securely and never exposed.
      operationId: createAuthor
      security:
        - GeneralApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - keyType
              properties:
                name:
                  type: string
                  description: Name of the author
                description:
                  type: string
                  description: Description of the author
                url:
                  type: string
                  format: uri
                  description: URL associated with the author
                keyType:
                  type: string
                  enum: [HUMAN, AI, HUMAN_AI_MIX, ORGANIZATION]
                  description: Type of the author key
                keyAlgorithm:
                  type: string
                  enum: [RSA, ECDSA, ED25519]
                  default: RSA
                  description: The cryptographic algorithm to use for the key pair
      responses:
        "201":
          description: Author created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  author:
                    $ref: "#/components/schemas/Author"
                  authorApiKey:
                    type: string
                    description: API key for author-specific operations (only returned once at creation)
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags:
        - Authors
      summary: List authors
      description: Returns a list of authors with optional filtering
      operationId: listAuthors
      security:
        - GeneralApiKey: []
      parameters:
        - name: name
          in: query
          description: Filter by author name
          schema:
            type: string
        - name: keyType
          in: query
          description: Filter by key type
          schema:
            type: string
            enum: [HUMAN, AI, HUMAN_AI_MIX, ORGANIZATION]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of authors
          content:
            application/json:
              schema:
                type: object
                properties:
                  authors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Author"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /authors/{authorId}:
    get:
      tags:
        - Authors
      summary: Get author details
      description: Returns details about a specific author
      operationId: getAuthor
      parameters:
        - name: authorId
          in: path
          required: true
          description: ID of the author
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Author details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "404":
          description: Author not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Authors
      summary: Update author details
      description: Updates details for a specific author
      operationId: updateAuthor
      security:
        - AuthorApiKey: []
      parameters:
        - name: authorId
          in: path
          required: true
          description: ID of the author
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the author
                description:
                  type: string
                  description: Description of the author
                url:
                  type: string
                  format: uri
                  description: URL associated with the author
      responses:
        "200":
          description: Author updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - API key does not match author
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Author not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Authors
      summary: Delete an author
      description: Deletes an author and associated keys
      operationId: deleteAuthor
      security:
        - AuthorApiKey: []
      parameters:
        - name: authorId
          in: path
          required: true
          description: ID of the author
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Author deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - API key does not match author
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Author not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /authors/{authorId}/public-key:
    get:
      tags:
        - Authors
      summary: Get author's public key
      description: Returns the public key for a specific author
      operationId: getAuthorPublicKey
      parameters:
        - name: authorId
          in: path
          required: true
          description: ID of the author
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Author's public key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKey"
        "404":
          description: Author or public key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /claims:
    post:
      tags:
        - Claims
      summary: Create a new claim type
      description: Creates a new claim type definition
      operationId: createClaimType
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
              properties:
                name:
                  type: string
                  description: Name of the claim type
                description:
                  type: string
                  description: Description of what this claim type represents
                possibleValues:
                  type: array
                  description: Possible values for this claim type (if applicable)
                  items:
                    type: string
      responses:
        "201":
          description: Claim type created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags:
        - Claims
      summary: List claim types
      description: Returns a list of claim types
      operationId: listClaimTypes
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of claim types
          content:
            application/json:
              schema:
                type: object
                properties:
                  claims:
                    type: array
                    items:
                      $ref: "#/components/schemas/Claim"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
  /claims/{claimId}:
    get:
      tags:
        - Claims
      summary: Get claim type details
      description: Returns details about a specific claim type
      operationId: getClaimType
      parameters:
        - name: claimId
          in: path
          required: true
          description: ID of the claim type
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Claim type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          description: Claim type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...